requirements.txt
# Core dependencies
ccxt>=4.0.0
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Logging & monitoring (optional)
colorama>=0.4.6  # Colored terminal output
